Add support for systemd
In order to synchronize our service launch on boot we need to add explicit support for
systemd to notify systemd when we're really set up and other depending services can be
launched. As ls-hubd is a critical component of the system it has to be really up before
anyone can use it in order not to fail.

Upstream-Status: pending (http://github.com/openwebos/luna-service2)

Open-webOS-DCO-1.0-Signed-off-by: Simon Busch <morphis@gravedo.de>